refactor(proposer): extract proof submitter module from pipeline
Add a new `base_proposer::proof_submitter` module exposing
`ProofSubmitter` and `SubmitAction`. The submitter owns the
`OutputProposer`, the rollup client, the L1 client, the proposer address,
the block / intermediate block intervals, the TEE image hash, the optional
TEE prover registry address, and the canonical-output fetch concurrency.

`ProofSubmitter::submit` performs the full pre-flight pipeline:
1. JIT validation: re-fetches the canonical output root from the rollup
   node and compares it against the aggregate proposal root.
2. Intermediate-root validation: extracts per-block intermediate roots
   from the proposals, fetches fresh canonical roots for each non-target
   intermediate block, and verifies all match the canonical chain.
3. Optional TEE signer validation against the on-chain
   `TEEProverRegistry`, falling through on RPC failure so that a
   misconfigured registry doesn't permanently block submission.
4. Calls `OutputProposer::propose_output` and maps contract-level errors
   into `SubmitAction::{GameAlreadyExists, Discard, Failed}`.

The pipeline retains the `submit_tasks` JoinSet, `try_submit`,
`handle_submit_result`, and the `SubmitOutcome` cross-task channel — these
are coordinator concerns. `validate_and_submit` becomes a thin pipeline
wrapper that delegates to `ProofSubmitter::submit`, so all existing tests
keep their existing call sites.

Pipeline-side cleanup:
- The `fetch_canonical_roots_with(blocks, bypass_cache)` and
  `fetch_fresh_canonical_roots` helpers are gone; the only remaining
  pipeline caller is the recovery walk and proof-request build, which
  always use the cached path. The submitter has its own
  `fetch_fresh_canonical_roots` over its own rollup-client clone.
- The duplicate `SubmitAction` enum and its `Display` impl are removed
  from pipeline.rs and re-exported from lib.rs via the new module.
- A small `intermediate_block_numbers` helper remains on the pipeline so
  `forward_walk` can build `extraData` for UUID-based `games()`
  lookups; the submitter has its own equivalent driven by
  `PipelineConfig` values cached at construction time.

This is the second half of P6 from the proof-system consolidation plan;
P7 will remove the remaining transitional `ProverClient` adapter and the
dead `prove_tasks` JoinSet from the proposer pipeline.
