+++ crates/proof/proposer/src/lib.rs
@@ -28,6 +28,9 @@ pub use proof_adapter::{
 mod proof_collector;
 pub use proof_collector::{CollectedProof, ProofCollector};
 
+mod proof_submitter;
+pub use proof_submitter::{ProofSubmitter, SubmitAction};
+
 mod driver;
 pub use driver::{DriverConfig, PipelineHandle, ProposerDriverControl, RecoveredState};
